+++ FateConnect/Web/src/pages/Menu/styles.ts
@@ -72,6 +72,6 @@ export const IconDisc = styled(PolymorphicStack)(({ theme }) => ({
 
   '& svg': {
     fontSize: `${iconSizeTokens.lg}px`,
-    color: theme.palette.common.white,
+    color: theme.palette.secondary.contrastText,
   },
 }));